Refactor test fixtures
