+++ libsource/src/test/java/com/virtusize/libsource/data/parsers/ProductSizeJsonParserTest.kt
@@ -1,7 +1,7 @@
 package com.virtusize.libsource.data.parsers
 
 import com.google.common.truth.Truth.assertThat
-import com.virtusize.libsource.TestFixtures
+import com.virtusize.libsource.fixtures.TestFixtures
 import com.virtusize.libsource.data.remote.Measurement
 import com.virtusize.libsource.data.remote.ProductSize
 import org.json.JSONObject
