+++ libsource/src/test/java/com/virtusize/libsource/data/parsers/StoreJsonParserTest.kt
@@ -2,7 +2,7 @@ package com.virtusize.libsource.data.parsers
 
 import com.google.common.truth.Truth.assertThat
 import com.virtusize.libsource.data.remote.Store
-import com.virtusize.libsource.TestFixtures
+import com.virtusize.libsource.fixtures.TestFixtures
 import org.junit.Test
