+++ libsource/src/test/java/com/virtusize/libsource/data/remote/ProductTest.kt
@@ -5,7 +5,8 @@ import android.os.Build
 import androidx.test.core.app.ApplicationProvider
 import com.google.common.truth.Truth.assertThat
 import com.virtusize.libsource.R
-import com.virtusize.libsource.TestFixtures
+import com.virtusize.libsource.fixtures.ProductFixtures
+import com.virtusize.libsource.fixtures.TestFixtures
 import org.junit.Before
 import org.junit.Test
 
@@ -32,17 +33,17 @@ class ProductTest {
     @Test
     fun getRecommendationText_productIsAnAccessory_returnDefaultAccessoryText() {
         val defaultAccessoryText = context.getString(R.string.inpage_default_accessory_text)
-        assertThat(TestFixtures.storeProduct(18).getRecommendationText(i18nLocalization)).isEqualTo(defaultAccessoryText)
-        assertThat(TestFixtures.storeProduct(19).getRecommendationText(i18nLocalization)).isEqualTo(defaultAccessoryText)
-        assertThat(TestFixtures.storeProduct(25).getRecommendationText(i18nLocalization)).isEqualTo(defaultAccessoryText)
-        assertThat(TestFixtures.storeProduct(26).getRecommendationText(i18nLocalization)).isEqualTo(defaultAccessoryText)
+        assertThat(ProductFixtures.storeProduct(18).getRecommendationText(i18nLocalization)).isEqualTo(defaultAccessoryText)
+        assertThat(ProductFixtures.storeProduct(19).getRecommendationText(i18nLocalization)).isEqualTo(defaultAccessoryText)
+        assertThat(ProductFixtures.storeProduct(25).getRecommendationText(i18nLocalization)).isEqualTo(defaultAccessoryText)
+        assertThat(ProductFixtures.storeProduct(26).getRecommendationText(i18nLocalization)).isEqualTo(defaultAccessoryText)
     }
 
     @Test
     fun getRecommendationText_productIsNotAnAccessory_returnNoDataText() {
         val noDataText = context.getString(R.string.inpage_no_data_text)
-        assertThat(TestFixtures.storeProduct(4).getRecommendationText(i18nLocalization)).isEqualTo(noDataText)
-        assertThat(TestFixtures.storeProduct(7).getRecommendationText(i18nLocalization)).isEqualTo(noDataText)
-        assertThat(TestFixtures.storeProduct(15).getRecommendationText(i18nLocalization)).isEqualTo(noDataText)
+        assertThat(ProductFixtures.storeProduct(4).getRecommendationText(i18nLocalization)).isEqualTo(noDataText)
+        assertThat(ProductFixtures.storeProduct(7).getRecommendationText(i18nLocalization)).isEqualTo(noDataText)
+        assertThat(ProductFixtures.storeProduct(15).getRecommendationText(i18nLocalization)).isEqualTo(noDataText)
     }
 }
